+++ js/main-js.js
@@ -1,7 +1,7 @@
-import { mainDiv, mainForm, checkedList } from "./jsModuleObg.js";
-import { htmlBody } from "./storage.js";
-import { header } from "./headerStructure.js";
+import { mainDiv, mainForm, checkedList } from "../todoForm/inputListForm.js";
+import { htmlBody } from "../storage/storage.js";
+import { header } from "../topPanel/headerStructure.js";
 mainDiv.prepend(mainForm);
 mainDiv.prepend(checkedList);
-htmlBody.prepend(header)
-console.log(htmlBody)
+htmlBody.prepend(header);
+console.log(htmlBody);
